Add support for compiling WebEngine with Clang sanitizer options.
This change pipes the already supported Qt sanitize options to
Chromium, so the Chromium build system adds the correct compiler
sanitizer flags.

The change also checks, whether the requirements to build a sanitizer
enabled Chromium, are met (currently works properly on OS X with a
non-apple clang, and on Linux with clang + libc++).

Also because of a limitation in how mac-clang mkspec sets the
Apple compilers, it is not possible to change compilers even in a
custom mkspec. So what happens is that Qt is compiled with Apple
clang, and WebEngine is compiled with a custom non-Apple clang found
in the path.

This is a cherry-pick of https://codereview.qt-project.org/r/155225.

Conflicts:
	src/core/config/common.pri
	src/core/config/desktop_linux.pri
	src/core/config/mac_osx.pri
	src/process/process.pro

Change-Id: Ibab2d5735935ddc92226575fc0e9f5912b807329
Task-number: QTBUG-52656